script: move media interfaces to script/dom/media module
