+++ components/script/dom/media/mediadevices.rs
@@ -21,9 +21,9 @@ use crate::dom::bindings::reflector::{DomGlobal, reflect_dom_object};
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::eventtarget::EventTarget;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::mediadeviceinfo::MediaDeviceInfo;
-use crate::dom::mediastream::MediaStream;
-use crate::dom::mediastreamtrack::MediaStreamTrack;
+use crate::dom::media::mediadeviceinfo::MediaDeviceInfo;
+use crate::dom::media::mediastream::MediaStream;
+use crate::dom::media::mediastreamtrack::MediaStreamTrack;
 use crate::dom::promise::Promise;
 use crate::realms::{AlreadyInRealm, InRealm};
 use crate::script_runtime::CanGc;
