+++ components/script/dom/media/mod.rs
@@ -1,0 +1,15 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+pub(crate) mod mediadeviceinfo;
+pub(crate) mod mediadevices;
+pub(crate) mod mediaerror;
+pub(crate) mod mediafragmentparser;
+pub(crate) mod medialist;
+pub(crate) mod mediametadata;
+pub(crate) mod mediaquerylist;
+pub(crate) mod mediaquerylistevent;
+pub(crate) mod mediasession;
+pub(crate) mod mediastream;
+pub(crate) mod mediastreamtrack;
