+++ components/script/dom/mod.rs
@@ -323,17 +323,8 @@ pub(crate) mod intersectionobserver;
 pub(crate) mod intersectionobserverentry;
 pub(crate) mod keyboardevent;
 pub(crate) mod location;
-pub(crate) mod mediadeviceinfo;
-pub(crate) mod mediadevices;
-pub(crate) mod mediaerror;
-pub(crate) mod mediafragmentparser;
-pub(crate) mod medialist;
-pub(crate) mod mediametadata;
-pub(crate) mod mediaquerylist;
-pub(crate) mod mediaquerylistevent;
-pub(crate) mod mediasession;
-pub(crate) mod mediastream;
-pub(crate) mod mediastreamtrack;
+pub(crate) mod media;
+pub(crate) use self::media::*;
 pub(crate) mod messagechannel;
 pub(crate) mod messageevent;
 #[allow(dead_code)]
